add backup script

+++ python/utils/files.py
@@ -8,6 +8,10 @@ from pathlib import Path
 CIPHER = "AES256"
 
 
+class FileAlreadyExistsError(Exception):
+    pass
+
+
 def compare_files(src_file, dst_file):
     src_hash = hash_file(src_file)
     dst_hash = hash_file(dst_file)
@@ -16,12 +20,15 @@ def compare_files(src_file, dst_file):
         logging.error(f"{dst_file} doesn't match {src_file}")
         return False
 
-    logging.error(f"{dst_file} and {src_file} are the same files")
+    logging.info(f"{dst_file} and {src_file} are the same files")
     return True
 
 
 def create_checksum_file(file):
     checksum_file = f"{file}.sha256"
+
+    if os.path.exists(checksum_file):
+        raise FileAlreadyExistsError(f"{checksum_file} file already exists")
 
     with open(checksum_file, "w") as f:
         f.write(f"{hash_file(file)} {file}\n")
@@ -30,8 +37,13 @@ def create_checksum_file(file):
     return checksum_file
 
 
-def encrypt_file(file, passphrase):
-    encrypted_file = f"{file}.gpg"
+def encrypt_file(file, passphrase, encrypted_file=None):
+    if encrypted_file is None:
+        encrypted_file = f"{file}.gpg"
+
+    if os.path.exists(encrypted_file):
+        raise FileAlreadyExistsError(f"{encrypted_file} file already exists")
+
     gpg = gnupg.GPG()
 
     with open(file, "rb") as f:
@@ -47,13 +59,18 @@ def encrypt_file(file, passphrase):
     return encrypted_file
 
 
-def decrypt_file(file, passphrase):
+def decrypt_file(file, passphrase, decrypted_file=None):
     file_path = Path(file)
 
     if file_path.suffix != ".gpg":
-        raise Exception("ERROR: only files with gpg extension are supported")
+        raise Exception("Only files with gpg extension are supported")
 
-    decrypted_file = str(file_path.with_suffix(""))
+    if decrypted_file is None:
+        decrypted_file = str(file_path.with_suffix(""))
+
+    if os.path.exists(decrypted_file):
+        raise FileAlreadyExistsError(f"{decrypted_file} file already exists")
+
     gpg = gnupg.GPG()
 
     with open(file, "rb") as f:
@@ -88,10 +105,15 @@ def rename_file(src_pattern, dst):
     if len(files) > 1:
         raise ValueError()
 
+    if os.path.exists(dst):
+        raise FileAlreadyExistsError(f"{dst} file already exists")
+
     os.rename(files[0], dst)
     logging.info(f"File {files[0]} renamed to {dst}")
 
 
 def shred_file(file):
+    if not os.path.exists(file):
+        raise FileNotFoundError()
     os.system(f"shred -u {file}")
     logging.info(f"File {file} wiped")
